+++ handlers/element/delete.php
@@ -7,14 +7,16 @@ if(!isset($_SESSION["name"])){
     if(!$_SESSION['edit']){
         header('Location: ../../index.php?board='.$_SESSION['board']);
     }else{
+        if(is_numeric($_GET["id"])){
 
-        include '../../src/cfg/conn.php';
-
-        $sql = "DELETE FROM `element` WHERE `element`.`id_element` = ".$_GET['id'];
-
-        $res = @mysqli_query($conn, $sql);
-
-        mysqli_close($conn);
+            include '../../src/cfg/conn.php';
+    
+            $sql = "DELETE FROM `element` WHERE `element`.`id_element` = ".$_GET['id'];
+    
+            $res = @mysqli_query($conn, $sql);
+    
+            mysqli_close($conn);
+        }
 
         header('Location: ../../index.php?board='.$_SESSION['board']);
     }
